Add files via upload
In this project, I build a smart beta portfolio and compare it to a benchmark index. To find out how well the smart beta portfolio did, I calculate the tracking error against the index. I then build a portfolio by using quadratic programming to optimize the weights. The code will rebalance this portfolio and calculate turn over to evaluate the performance. I use this metric to find the optimal rebalancing Frequency. For the dataset, I have been using the end of day from Quotemedia.